move view models over to domain
