+++ src/domain/session/roomlist/RoomTileViewModel.js
